added Client Request/Stub object definitions, corrected other client application object definitions
Created file to define the Client Request Object (a stub object that will represent a full job submission process for one job requested by the client).  Also modified the files for the Job Callback Handler and Client Logger objects to ensure, e.g., all resources are closed for cleanup once a job is completed or cancelled, and so that all variables/fields in the Job Callback Object would be used.