changed client log filepath variable in NewClientLoggingObject()
In the NewClientLoggingObject() constructor function, the clientLogFilename variable was changed to include not only creating .log files for client activity logs, but other formats (e.g., .txt) as well.  All that matters is that for right now, the root directory path for all the client activity logs in the client app will be 'jobmgr/client/clientlogs', from where users of the app can also create separate folders within (if needed) to organize different client logs, or just create a new log file or use an existing log file right from this root directory.