created CLI object
I created a CLI object, which will parse the flags of the  command and create a single ClientRequestObject to request that the specified job and its process are executed by the Job Manager Service.  This object is created to try and demonstrate how the core ClientRequestObject can be implemented in different ways, where all that a user would need to do, for each job they want to run, is: 1) create a new ClientRequestObject for each job; 2) connect each client to the Job Manager Server using its ConnectToServer() method; 3) submit and monitor each job by calling each corresponding client's SubmitJob() method.