Added Individual Job Handler and components
Before I add the functionality for the Job Executor Service and the components for the general Job Manager Service server, I needed to define the Job Handler and all the methods and components needed to submit and run the individual job.  The initial JobHandler definition contains the implemented RPCs of the jobmgrcapnp.JobQueryCancel service interface, because ultimately the job handled by this handler will be controlled by the client calling on either the queryJob RPC to monitor the current active/inactive status of the job, or the cancelJob RPC to request that the job be cancelled.

Then, I defined 'helper' methods that users of the Job Manager Server application can use when defining their own processes to run, so that they need not carry the burden of needing to repeat different server and client logging calls, or directly call the associated job callback's RPCs, in case an error occurs during any of those RPC calls -- the helper will handle all of that for them.

Finally, I defined a simple example process that the server can try running, and also created the Process Choices Map to register that method, so that the Job Executor service can choose the appropriate process from that map to run on behalf of the requesting client from the service's client application.