Added General Server components
Now that I have added all the service definitions (the Individual Job Handler and the Job Executor services) and their RPCs or internal methods, I added in the definition of the General Server components for setting up the entire Job Manager Service.  The General Server has 3 basic components: a constructor NewJobManagerServer function, a ListenAndServe method for setting up and serving client requests, and a Shutdown method for releasing the root capability and closing the network listener, so that then the ListenAndServe method can close the attached ServerLoggingObject resources for proper cleanup and prevention of resource leaks.