modified .gitignore to include process results folder
One of the last cleanup tasks I decided to do before merging the server-app branch changes to the Job Manager Service is add the relative path to server/jobresults/ which is a directory that users would create and put on their Job Manager Service server app instance to collect all the results of running processes.  This is for security purposes -- so that potential threat actors and attackers cannot view any results of what a server instance has run.