Able to modify return type of job executor and job handlers to pointer
In the definitions of the Job Executor and Job Handler RPCs, I saw that the return type of 'JobExecutorHandler' and 'JobHandler' were raising warnings because they were using a sync.Mutex.  When I changed the return types of the RPC methods to '*JobExecutorHandler' and '*JobHandler', respectively, the warnings were removed, and still the pointer return type satisfied the interfaces of the original Cap'n Proto Go implementations of the jobmgrcapnp.JobExecutor and jobmgrcapnp.JobQueryCancel interfaces.