Added functionality for storing log and process result files in appropriate directories
I have been able to easily add in functionality that allows the client and server programs to change directories for each of the client and server activity logs, as well as the clients' individual job logs, and also the files for the process results on the server side.  The .gitiginore file for this service has also been updated to ensure those directories storing the logs and process results are ignored from version control.  It just requires that before a service user starts running client-server interaction sessions, that they set up a jobmgr/client/clientlogs directory, a jobmgr/client/clientlogs/joblogs directory, a jobmgr/server/serverlogs directory, and a jobmgr/server/processresults directory.